Create Node List from Model Expression

Creates a list of nodes based on a given model expression, handling input, output, and previous nodes with their properties and relationships.

Usage

get_node_list(
  exp,
  param_names = NULL,
  mctable = set_mctable(),
  data_keys = set_data_keys()
)

Arguments

exp

An R expression containing model calculations

param_names

Optional named vector for parameter renaming

mctable

Reference table for mcnodes, defaults to set_mctable()

data_keys

Data structure and keys, defaults to set_data_keys()

Value

A list of class "mcnode_list" containing node information

Examples

mcmodule:::get_node_list(
  exp = imports_exp,
  mctable = imports_mctable,
  data_keys = imports_data_keys)
#> $w_prev
#> $w_prev$type
#> [1] "in_node"
#> 
#> $w_prev$mc_func
#> [1] "runif"
#> 
#> $w_prev$description
#> [1] "Within herd prevalence"
#> 
#> $w_prev$inputs_col
#> [1] "w_prev_min" "w_prev_max"
#> 
#> $w_prev$input_dataset
#> [1] "prevalence_region"
#> 
#> $w_prev$keys
#> [1] "pathogen" "origin"  
#> 
#> $w_prev$module
#> [1] "imports"
#> 
#> $w_prev$mc_name
#> [1] "w_prev"
#> 
#> 
#> $test_origin
#> $test_origin$type
#> [1] "in_node"
#> 
#> $test_origin$description
#> [1] "Probability of the animals being tested in origin"
#> 
#> $test_origin$inputs_col
#> [1] "test_origin"
#> 
#> $test_origin$input_dataset
#> [1] "prevalence_region"
#> 
#> $test_origin$keys
#> [1] "pathogen" "origin"  
#> 
#> $test_origin$module
#> [1] "imports"
#> 
#> $test_origin$mc_name
#> [1] "test_origin"
#> 
#> 
#> $test_sensi
#> $test_sensi$type
#> [1] "in_node"
#> 
#> $test_sensi$mc_func
#> [1] "rpert"
#> 
#> $test_sensi$description
#> [1] "Test sensitivity"
#> 
#> $test_sensi$inputs_col
#> [1] "test_sensi_min"  "test_sensi_mode" "test_sensi_max" 
#> 
#> $test_sensi$input_dataset
#> [1] "test_sensitivity"
#> 
#> $test_sensi$keys
#> [1] "pathogen"
#> 
#> $test_sensi$module
#> [1] "imports"
#> 
#> $test_sensi$mc_name
#> [1] "test_sensi"
#> 
#> 
#> $inf_a
#> $inf_a$node_exp
#> [1] "w_prev"
#> 
#> $inf_a$type
#> [1] "out_node"
#> 
#> $inf_a$inputs
#> [1] "w_prev"
#> 
#> $inf_a$module
#> [1] "imports"
#> 
#> $inf_a$mc_name
#> [1] "inf_a"
#> 
#> $inf_a$keys
#> [1] "pathogen" "origin"  
#> 
#> 
#> $false_neg_a
#> $false_neg_a$node_exp
#> [1] "inf_a * test_origin * (1 - test_sensi)"
#> 
#> $false_neg_a$type
#> [1] "out_node"
#> 
#> $false_neg_a$inputs
#> [1] "inf_a"       "test_origin" "test_sensi" 
#> 
#> $false_neg_a$module
#> [1] "imports"
#> 
#> $false_neg_a$mc_name
#> [1] "false_neg_a"
#> 
#> $false_neg_a$keys
#> [1] "pathogen" "origin"  
#> 
#> 
#> $no_test_a
#> $no_test_a$node_exp
#> [1] "inf_a * (1 - test_origin)"
#> 
#> $no_test_a$type
#> [1] "out_node"
#> 
#> $no_test_a$inputs
#> [1] "inf_a"       "test_origin"
#> 
#> $no_test_a$module
#> [1] "imports"
#> 
#> $no_test_a$mc_name
#> [1] "no_test_a"
#> 
#> $no_test_a$keys
#> [1] "pathogen" "origin"  
#> 
#> 
#> $no_detect_a
#> $no_detect_a$node_exp
#> [1] "false_neg_a + no_test_a"
#> 
#> $no_detect_a$type
#> [1] "out_node"
#> 
#> $no_detect_a$inputs
#> [1] "false_neg_a" "no_test_a"  
#> 
#> $no_detect_a$module
#> [1] "imports"
#> 
#> $no_detect_a$mc_name
#> [1] "no_detect_a"
#> 
#> $no_detect_a$keys
#> [1] "pathogen" "origin"  
#> 
#> 
#> attr(,"class")
#> [1] "mcnode_list"